requirements.txt
flask
playwright
pandas
openpyxl
python-dotenv
requests
beautifulsoup4
lxml
